Debugged warning on ROM loading

+++ rom_load.h
@@ -1,6 +1,8 @@
 #ifndef LOAD_ROM
 #define LOAD_ROM
 
-void load_rom(char* filename, int* data);
+#include "utils.h"
+
+void load_rom(char* filename, BYTE* data);
 
 #endif
